* gc.c (ruby_xmalloc): use size_t for malloc argument instead of long.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@16112 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ gc.c
@@ -314,7 +314,7 @@ ruby_xmalloc(size_t size)
 void *
 ruby_xmalloc2(size_t n, size_t size)
 {
-    long len = size * n;
+    size_t len = size * n;
     if (n != 0 && size != len / n) {
 	rb_raise(rb_eArgError, "malloc: possible integer overflow");
     }
